Allow body parameter type name to be customized
The only DataType value set outside of the swagger model builder is the
result of calling `.Reads(...)` on a Route. In order to allow someone to
properly generate a consistent Swagger doc where the default
`reflect.TypeOf().String()` call would not be unique across an API for
multiple types, add a new method to WebService to allow the caller to
define their own pattern for this.