Add support for route condition functions
Sometimes you need to be able to say that a route only
applies under certain conditions. Now you can include
`.If(RouteSelectionConditionFunction)` to your route
definition to add a function that will be called during
route selection to help determine whether the route
is applicable to the incoming HTTP request.