log: Deprecate in favor of github.com/go-log/log
go-restful's Logger interface is from b80aca09 (use an interface for
Logger instead of concrete type, 2015-03-16, #188), at which point the
interface included both the Print and Fatal method families.  That
interface moved to the log subpackage in 18c1d52a (Add 'log' package
for package-wide logging, expose SetLogger(Logger), 2015-03-18, #192).
And the interface was reduced to just Print and Printf in a40a6af9
(Reduce StdLogger interface to minimum, conform callers,
2015-03-23, #194), but the interface is clearly not
go-restful-specific.

More recently, github.com/go-log/log [1,2] was created as a
stand-alone package with a similar interface (it has Log and Logf
instead of Print and Printf).  Just this morning, go-log grew a
subpackage to convert from Print/Printf to Log/Logf [3].  That means
we can offload our generic-logger shim to go-log, save a few lines of
generic code here (after removing the deprecated code), and
automatically benefit from go-logs other implementations (e.g. you can
use info.New [4] to plug in a logger based on glog's Verbose [5]).

I've deprecated the old interfaces, but kept them around to facilitate
migrations to the new approach.  The only breaking change should be
the removal of log.Logger, so that folks who are setting that directly
(vs. going through one of the two SetLogger functions) will get
compile-time errors and notice that they have to update their usage.
If we don't mind jumping through a few extra hoops, we could keep
log.Logger around for now and stage migration with:

Phase 1:

* Replace the public restful.Logger with a restful.SetLogLogger
  setter, that takes a go-log (Log/Logf) interface.
* Calling restful.SetLogger, restful.SetLogLogger, or log.SetLogger
  will all set log.Logger.  We'd need a Log/Logf -> Print/Printf shim
  for SetLogLogger.
* Keep importing go-restful/log and using log.Print and log.Printf
  where I'm currently using Logger.Log and Logger.Logf.

Phase 2:

* Drop go-restful/log and the other interfaces deprecated in this
  commit.
* Replace the removed log.Logger global with a new restful.Logger.
* Replace the log.Print and log.Printf calls with Logger.Log and
  Logger.Logf.
* Deprecate SetLogLogger.

Phase 3:

* Drop SetLogLogger.

Alternatively, you could make the global private (restful.logger) and
keep SetLogLogger.  But I expect the go-log Logger interface to be
stable enough that we don't need to hide it behind a setter,
expecially one with as ugly a name as SetLogLogger, or SetLogger2, or
whatever ;).

The .travis.yml addition avoids [6]:

  $ go test -v
  logger.go:10:2: cannot find package "github.com/go-log/log" in any of:
    /home/travis/.gimme/versions/go1.11beta1.linux.amd64/src/github.com/go-log/log (from $GOROOT)
    /home/travis/gopath/src/github.com/go-log/log (from $GOPATH)
  logger.go:11:2: cannot find package "github.com/go-log/log/print" in any of:
    /home/travis/.gimme/versions/go1.11beta1.linux.amd64/src/github.com/go-log/log/print (from $GOROOT)
    /home/travis/gopath/src/github.com/go-log/log/print (from $GOPATH)

Apparently this is the first external dep involved in the integration
tests (with json-iterator/go hidden behind the jsoniter build tag).

[1]: https://github.com/go-log/log
[2]: https://godoc.org/github.com/go-log/log
[3]: https://godoc.org/github.com/go-log/log/print#New
[4]: https://godoc.org/github.com/go-log/log/info#New
[5]: https://godoc.org/github.com/golang/glog#Verbose
[6]: https://travis-ci.org/emicklei/go-restful/builds/405930109#L477