Reduce allocations in per-request methods to improve performance (#395)
* Prevent additional allocations when sorting routes for selection

Use of an uninitialized array and the use of sort.Reverse cause
extra allocations in request routing which are unnecessary.

* Reduce allocations in Accept and Content-Type matching

The mime handling methods can do a single pass over the header without
needing to allocate. This improves performance of these methods in
higher traffic API servers.

* Reduce allocations in route selection

The use of new arrays for each phase of detectRoute can be simplified:
each pass filters the existing array and whatever is left at the end
is returned. While somewhat harder to read, this method is a significant
source of allocations.